backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
google-generativeai==0.8.4
youtube-transcript-api==1.1.1
langdetect==1.0.9
gunicorn==23.0.0
python-dotenv
